# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AjustarEstoqueRequest:
    properties:
      motivo:
        type: string
      nova_quantidade:
        description: Aceita decimal como string
        type: string
      produto_id:
        type: string
    required:
    - motivo
    - nova_quantidade
    - produto_id
    type: object
  dto.AlertaEstoqueBaixo:
    properties:
      categoria:
        type: string
      nome:
        type: string
      percentual:
        description: Percentual do estoque (atual/mínimo * 100)
        type: string
      produto_id:
        type: string
      quantidade_atual:
        type: string
      quantidade_minima:
        type: string
      severidade:
        description: crítico, alerta, baixo
        type: string
      sku:
        type: string
      unidade_medida:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.ItemEntradaRequest:
    properties:
      produto_id:
        type: string
      quantidade:
        type: integer
      valor_unitario:
        type: string
    required:
    - produto_id
    - quantidade
    - valor_unitario
    type: object
  dto.ListAlertasEstoqueBaixoResponse:
    properties:
      alertas:
        items:
          $ref: '#/definitions/dto.AlertaEstoqueBaixo'
        type: array
      total:
        type: integer
    type: object
...
  dto.MovimentacaoResponse:
    properties:
      created_at:
        type: string
      data:
        description: Data da movimentação
        type: string
      fornecedor_id:
        type: string
      id:
        type: string
      observacoes:
        type: string
      produto_id:
        type: string
      produto_nome:
        type: string
      quantidade:
        description: Decimal como string
        type: string
      tenant_id:
        type: string
      tipo:
        type: string
      usuario_id:
        type: string
      valor_total:
        type: string
      valor_unitario:
        type: string
    type: object
...
  dto.RegistrarEntradaRequest:
    properties:
      data_entrada:
        description: YYYY-MM-DD
        type: string
      fornecedor_id:
        type: string
      gerar_financeiro:
        description: Se true, cria conta a pagar
        type: boolean
      itens:
        items:
          $ref: '#/definitions/dto.ItemEntradaRequest'
        minItems: 1
        type: array
      observacoes:
        type: string
    required:
    - fornecedor_id
    - itens
    type: object
  dto.RegistrarSaidaRequest:
    properties:
      motivo:
        enum:
        - VENDA
        - USO_INTERNO
        - PERDA
        - DEVOLUCAO
        type: string
      observacoes:
        type: string
      produto_id:
        type: string
      quantidade:
        description: Aceita decimal como string
        type: string
    required:
    - motivo
    - produto_id
    - quantidade
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nexo.barber
    name: API Support
    url: https://nexo.barber/support
  description: API completa para gestão de barbearias com módulos de Agendamento,
    Comanda, Financeiro, Estoque e Analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://nexo.barber/terms
  title: NEXO - Barber Analytics API
  version: "2.0"
paths:
...
  /api/v1/stock/adjust:
    post:
      consumes:
      - application/json
      description: Realiza ajuste manual de quantidade com auditoria
      parameters:
      - description: Dados do ajuste
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AjustarEstoqueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovimentacaoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ajustar estoque manualmente
      tags:
      - Estoque
  /api/v1/stock/alerts:
    get:
      consumes:
      - application/json
      description: Lista produtos com estoque abaixo do minimo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListAlertasEstoqueBaixoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar alertas de estoque baixo
      tags:
      - Estoque
  /api/v1/stock/entries:
    post:
      consumes:
      - application/json
      description: Registra entrada de produtos no estoque (compra/recebimento)
      parameters:
      - description: Dados da entrada
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegistrarEntradaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar entrada de estoque
      tags:
      - Estoque
  /api/v1/stock/exit:
    post:
      consumes:
      - application/json
      description: Registra saida de produto (venda, consumo interno, perda)
      parameters:
      - description: Dados da saida
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegistrarSaidaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MovimentacaoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar saida de estoque
      tags:
      - Estoque
securityDefinitions:
  BearerAuth:
    description: 'Bearer token JWT (formato: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...